move weather forecast to separate utility
